Migrate most structs
