+++ bellows/ezsp/v5/commands.py
@@ -674,7 +674,7 @@ COMMANDS = {
     "getTransientLinkKey": (
         0xCE,
         (t.EUI64,),
-        (t.EmberStatus, t.EmberTransientKeyData),
+        (t.EmberStatus, t.EmberTransientKeyDataV5),
     ),
     # 10. Trust Center Frames
     "trustCenterJoinHandler": (
