+++ bellows/ezsp/v6/commands.py
@@ -793,7 +793,7 @@ COMMANDS = {
     "getTransientLinkKey": (
         0xCE,
         (t.EUI64,),
-        (t.EmberStatus, t.EmberTransientKeyData),
+        (t.EmberStatus, t.EmberTransientKeyDataV5),
     ),
     # 10. Trust Center Frames
     "trustCenterJoinHandler": (
